+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed/updated unknown extended profile keys rendering.
 - Added support for `---`, `-#`, and fixed `-` to be unordered.
+- Fix quotes around colors for tuwunel homeservers ([#46](https://github.com/7w1/sable/issues/46))
 
 ## 1.2.1